src/requirements.txt
bottle==0.12.13
bottle-fdsend==0.1.1
bottle-sqlite==0.1.3
click==6.7
beaker==1.9.0

# Testing
webtest==2.0.28
pytest==3.2.1
